fix(build,ci): FreeBSD build, and three CI steps that could never pass
FreeBSD: syscall.Statfs_t is platform-specific, and Bavail is uint64 on
Linux but int64 on the BSDs. `st.Bavail * uint64(st.Bsize)` compiles on
Linux and fails the FreeBSD build with a type mismatch. Both operands are
now converted. All six release targets cross-compile.

tools/bench: the coverage guard was an awk one-liner with an unterminated
quote, so the step printed the coverage and then exited 2 regardless of it.
It has never passed. Rewritten as a complete awk program, checked against
values either side of the floor.

mddb-chat: the build script runs prost-build, which shells out to protoc,
and the runner does not ship one -- so the job failed in the build script
and never reached a test. Added the install step.

gosec: a deterministic benchmark seed tripped G404 in internal/vector.
Silenced with the same annotation the repo's other benchmark fixtures use.
It was missed locally because golangci-lint had been run on the package
rather than ./...

+++ services/mddbd/persistence_check.go
@@ -161,7 +161,13 @@ func freeBytes(dir string) (uint64, error) {
 	if err := syscall.Statfs(dir, &st); err != nil {
 		return 0, err
 	}
-	return st.Bavail * uint64(st.Bsize), nil // #nosec G115 -- block size is positive
+	// Both fields are converted rather than one: syscall.Statfs_t is
+	// platform-specific, and Bavail is uint64 on Linux but int64 on the BSDs.
+	// Multiplying them directly compiles on Linux and fails the FreeBSD build
+	// with a type mismatch, which is how this reached a release branch — the
+	// only builder that could see it is the one nobody runs locally.
+	// #nosec G115 -- a block count and a block size are both non-negative
+	return uint64(st.Bavail) * uint64(st.Bsize), nil
 }
 
 // logPersistenceStatus reports the findings at startup. Warnings are logged
